refactor: refactor app after retro
