+++ src/components/AppStartPage/Button.jsx
@@ -5,6 +5,7 @@ const StyledButton = styled.button`
     padding: 1rem 2rem;
     cursor: pointer;
     background: #383737;
+    color: white;
 `;
 
 const Button = ({ children }) => {
